refactor: Added CTA, Footer, Navbar in components - home folder
